+++ packages/sources/tradinghours/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @chainlink/tradinghours-adapter
+
+## 0.1.0
+
+### Minor Changes
+
+- [#3384](https://github.com/smartcontractkit/external-adapters-js/pull/3384) [`e01c723`](https://github.com/smartcontractkit/external-adapters-js/commit/e01c723dff9877866b779bcac9de32bebf1117b3) Thanks [@martin-cll](https://github.com/martin-cll)! - New adapter with market-status endpoint